Add a note about the known limitations of the credential claim metadata
As discussed on yesterday's working group call:

https://github.com/openid/OpenID4VCI/issues/266#issuecomment-1967533918

As it seems likely that https://github.com/openid/OpenID4VCI/pull/276
will not make it into implementer's draft 1, we instead add warnings
about the known limitations of the current data structure so that
implementer's are aware.

Note that I have not applied this to the mdoc profile section, as that
sections seems not to allow nested objects in the metadata.