Unify authorization scheme Bearer spelling in examples to "Bearer";
Although the authorization scheme is case-insensitive, all the examples
in the referred specifications RFC 6749 (and RFC 6750 referred by it)
use the spelling "Bearer", which is also the spelling in the IANA HTTP
Authentication Scheme Registry. Unified spelling reduces confusion and
possible mistakes in implementations that have missed the case-
insensitivity of the authentication scheme.