Fix GitHub action that publishes compiled version spec from main (#390)
Upgrade the download-artifact action from the version that github
blocked (due to a security issue) to latest, but that breaks other
things. So upgrade various gitlab actions to latest versions to
resolve this and other (non-fatal) warnings we were getting in the
frontend, e.g. things like:

The following artifacts were uploaded using a version of actions/upload-artifac\
t that is scheduled for deprecation: "github-pages".