Various arrays can't sensibly be empty, add "non-empty" to definitons
We already state that various arrays are optional or required, so
in both cases if they're actually present they need to have at least one
element. This avoids weird interoperability problems and means that
conformance testing can be simpler as it removes a case we'd otherwise
need to test for.

Similar to issue in VP: https://github.com/openid/OpenID4VP/issues/572