Cleanup language around c_nonce
We still had text that assumed the c_nonce was proactively sent to the wallet,
rather than (as is now the case) fetched by the wallet.

Also clarify that the nonce is a challenge rather than a nonce (without
changing any parameter names!), and that 'unpredictable' does not mean
that the same value cannot be used for a short time.

Some of the language is quite similar to the language in DPoP around the
DPoP nonces, as there is a lot of similarity - both mechanisms do not (fully)
prevent replay but do ensure freshness.

relates to some of the points raised in #461

relates the confusion that prompted #496

adds some clarify around the 'what nonces are' part of #491