Align VCI metadata retrieval with RFC8615 and OAuth Server Metadata
This means VCI uses the same algorithm for forming the .well-known path
for both the authorization server and the credential issuer.

This is likely to be necessary for the IANA registration for
openid-credential-issuer to succeed, and is also as a result of feedback
from VCI interop participants that having two different methods is
confusing.

Also fix incorrect link to the IANA Well-Known URIs registry.

closes #526