Attempt to make OAuth2 security recommendations more actionable
Daniel Fett pointed out that the current wording around the Best Current
Practice for OAuth 2.0 Security document is misleading as VCI does not
mention all the best current practices.

Reword to try and address this.

Add a recommendation to use FAPI2 as a way to comply with the BCP, and
add some notes around where FAPI2 might not be applicable.

Also replace the references to RFC9700 with BCP240 as it is likely to be
an evolving document.

closes #291