Make text about using credential_identifiers with scopes self-consistent
The text sometimes refers to a top level credential_identifiers in the
token response, and sometimes refers to it being within authorization_details.

As Stefan Charsley pointed out, there was a clear intent for it to be
within authorization_details stated here:

https://github.com/openid/OpenID4VCI/pull/392#issuecomment-2478638816

so fix the text that implies it is at the top level of the token response.

closes #514