Add "non-empty" to a few arrays added since the last pass
We already state that various arrays are optional or required, so in
both cases if they're actually present they need to have at least one
element. This avoids weird interoperability problems and means that
conformance testing can be simpler as it removes a case we'd otherwise
need to test for.

A follow up to https://github.com/openid/OpenID4VCI/pull/488 basically.