Remove Interactive Authorization Endpoint for now (#633)
Merging are per explicit agreement on Thursday's EU call that we'd include this in the version that went for voting.

* Remove Interactive Authorization Endpoint for now

As discussed on today's WG call, as documented here:

https://github.com/openid/OpenID4VCI/issues/595#issuecomment-3188933744

It's feeling a bit too rush and not feeling good enough to ship yet,
could be a footgun. Many people don't have a good feeling about it.

We'll look at how to re-introduce this separately, the main options being
doing a relatively swift 1.1 version, or creating a separate specification.

* Add missing document history entries