Move Encrypted Messages to the top level
Resolve #638

It wasn't clear from the comment as to which way the working group wished to resolve this. Moved it to the top level in this PR, mostly so that it could also be referenced by other sections (such as IAR) in the future, but happy to move it instead if people prefer.