Changes to update 1.1 spec to be a 1.1 draft (#660)
Merging as per previous agreement that chairs/editors would proceed with this, and it was discussed again on tonight's WG call and everyone seemed okay with it.

* Changes to update 1.1 spec to be a 1.1 draft

Change version, change title, add doc history back, remove text about erratas,
update github action to publish 1.1 (instead of 1.0).

* Move specs & examples into 1.0 and 1.1 subdirectories

As the examples are specific to a version this probably seems cleanest
and clearest.