Fixes #604 (#615)
Now that Paul has approved, merging as per discussion on today's WG call 

Additionally, and as an editorial change, I added interactive_authorization_endpoint to the AS metadata section, as it was previously missing. I also moved the corresponding normative "SHOULD use the IAE" statement into the AS metadata section, for consistency with how other endpoints are handled.

Co-authored-by: Joseph Heenan <joseph@heenan.me.uk>